MDEV-38913: pam test - separate config and module installations
MDEV-38913 handles the install of pam_mariadb_mtr directly by installing
it into /usr/lib/mysql/plugins/test_pam_modules in the mariadb-test-data
pacakge.

As such this testing script no longer needs to install this file.
Until this is merged up the previous functionality neesd to remain.

The test configuration file, mariadb_mtr, that has the included path
to the module and still needs to be installed.

+++ scripts/pam-test.sh
@@ -88,6 +88,8 @@ if [[ -f suite/plugins/pam/pam_mariadb_mtr.so ]]; then
   for p in /lib*/security /lib*/*/security ; do
     [[ -f "$p/pam_unix.so" ]] && sudo cp -v suite/plugins/pam/pam_mariadb_mtr.so "$p"/
   done
+fi
+if [[ -f suite/plugins/pam/mariadb_mtr ]]; then
   sudo cp -v suite/plugins/pam/mariadb_mtr /etc/pam.d/
 fi
